+++ src/PolarShadow.Services.Abstractions/Interfaces/IHistoryService.cs
@@ -8,9 +8,9 @@ namespace PolarShadow.Services
     public interface IHistoryService : ISyncAble<HistoryModel>
     {
         Task<ICollection<HistoryModel>> GetListPageAsync(int page, int pageSize, string filter = default);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
         Task AddOrUpdateAsync(HistoryModel model);
-        Task<HistoryModel> GetByIdAsync(int id);
+        Task<HistoryModel> GetByIdAsync(long id);
         Task<HistoryModel> GetByResourceNameAsync(string resourceName);
     }
 }
